Move showItemInFolderHandler to report export ipc
